getAccessToken function passes accessToken to App.jsx component where we use ternary operator to check if we're logged in already or not
