+++ server/server.js
@@ -79,8 +79,7 @@ app.get("/callback", (req, res) => {
       ).toString("base64")}`,
     },
   })
-    //Use access token to request data from Spotify API
-    //Destructuring the response.data to use in my queryParams, passing auth token to our front end
+    //Destructuring the response.data to pass auth token from BE to our FE in queryParams
     .then((response) => {
       if (response.status === 200) {
         const { access_token, refresh_token } = response.data;
